Rethink safety
After a few discussions with a few people, I think the best way to
handle safety in this crate is to have the user understand that Frida is
inherently an unsafe tool. So, you just have to use `unsafe` on the
`::obtain` function and then the world is your oyster.