frida-gum-sys,frida-sys: bindgen w/ prettyplease
The generated bindings are otherwise a single line which is hard for
vscode to open. It's useful to poke around in the bindings when playing
with the APIs. Prettyplease is fast. I don't see a downside to always
enabling it, and it was already enabled as a default feature of bindgen.